backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
websockets==12.0
spoon-ai-sdk
anthropic
termcolor
tenacity
pinecone
web3
solana
nest_asyncio
